+++ src/Users/ClientFactory.php
@@ -1,0 +1,33 @@
+<?php
+
+/**
+ * Vonage Client Library for PHP
+ *
+ * @copyright Copyright (c) 2016-2022 Vonage, Inc. (http://vonage.com)
+ * @license https://github.com/Vonage/vonage-php-sdk-core/blob/master/LICENSE.txt Apache License 2.0
+ */
+
+declare(strict_types=1);
+
+namespace Vonage\Users;
+
+use Psr\Container\ContainerInterface;
+use Vonage\Application\Client;
+use Vonage\Application\Hydrator;
+use Vonage\Client\APIResource;
+use Vonage\Client\Credentials\Handler\BasicHandler;
+
+class ClientFactory
+{
+    public function __invoke(ContainerInterface $container): Client
+    {
+        /** @var APIResource $api */
+        $api = $container->make(APIResource::class);
+        $api
+            ->setBaseUri('/v1/users')
+            ->setCollectionName('users')
+            ->setAuthHandler(new BasicHandler());
+
+        return new Client($api, new Hydrator());
+    }
+}
